Implementation of AI started. Added few services and methods to work with AI. Added new variable to Client_data table with AI points. Added new library to the project to work with the string comparison.

+++ src/main/java/web/dao/ClientDataDb.java
@@ -11,7 +11,7 @@ public class ClientDataDb {
     private int id;
     private ClientDb clientByIdClient;
     private FilmDb filmByIdFilm;
-    private int aiPoints;
+    private double aiPoints;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -62,12 +62,12 @@ public class ClientDataDb {
     }
 
     @Basic
-    @Column(name = "ai_points", nullable = false)
-    public int getAiPoints() {
+    @Column(name = "ai_points", nullable = false, precision = 0)
+    public double getAiPoints() {
         return aiPoints;
     }
 
-    public void setAiPoints(int aiPoints) {
+    public void setAiPoints(double aiPoints) {
         this.aiPoints = aiPoints;
     }
